Use correct types

+++ space invaders game/Assets/Spawn_Enemies.cs
@@ -10,7 +10,7 @@ public class Spawn_Enemies : MonoBehaviour
     public int enemies = 4;
     public List<GameObject> clones = new List<GameObject>();
 
-    private float padding = 2;
+    private int padding = 2;
     private int enemiesBefore;
     private bool spawned = false;
 
@@ -41,9 +41,9 @@ public class Spawn_Enemies : MonoBehaviour
 
             for (int i = 1; i < enemies + 1; i++)
             {
-                double offset = (i - (enemies + 1) * 0.5) * padding;
+                float offset = (i - (enemies + 1) * 0.5f) * padding;
                 GameObject clone;
-                clone = Instantiate(enemy, new Vector3(spawn.transform.position.x + ((float)offset), spawn.transform.position.y), enemy.transform.rotation);
+                clone = Instantiate(enemy, new Vector3(spawn.transform.position.x + offset, spawn.transform.position.y), enemy.transform.rotation);
                 clone.GetComponent<SpriteRenderer>().enabled = true;
                 clones.Add(clone);
             }
